Move corpus decode into fuzz/corpus.go, cmd into fuzz/cmd/seedcorpus
- Move fuzzInput type with *pprof.LabelsSnapshot and *pprof.ParseInput
  into corpus.go (package main), so fuzz.go just calls decodeFuzzInput
  and passes fields directly to ParseJFR
- Move seed corpus generation cmd from seedcorpus/cmd/ to cmd/seedcorpus/
- Delete seedcorpus/ package
